+++ src/main/java/com/swmStrong/demo/domain/leaderboard/service/LeaderboardService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface LeaderboardService {
-    void increaseScore(String category, String userId, long duration);
+    void increaseScore(String category, String userId, double duration);
     List<LeaderboardResponseDto> getTopUsers(String category, int topN);
     Optional<LeaderboardResponseDto> getUserScoreInfo(String category, String userId);
 }
